move vector methods to vector.ts
